backend/requirements.txt
Flask==2.1.1
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended
flask-mail